+++ apis/monitoring/v1alpha2/zz_alertpolicy_types.go
@@ -277,7 +277,7 @@ type ConditionMonitoringQueryLanguageTriggerParameters struct {
 	// that must fail the predicate for the
 	// condition to be triggered.
 	// +kubebuilder:validation:Optional
-	Count *int64 `json:"count,omitempty" tf:"count,omitempty"`
+	Count *float64 `json:"count,omitempty" tf:"count,omitempty"`
 
 	// The percentage of time series that
 	// must fail the predicate for the
@@ -505,7 +505,7 @@ type ConditionThresholdTriggerParameters struct {
 	// that must fail the predicate for the
 	// condition to be triggered.
 	// +kubebuilder:validation:Optional
-	Count *int64 `json:"count,omitempty" tf:"count,omitempty"`
+	Count *float64 `json:"count,omitempty" tf:"count,omitempty"`
 
 	// The percentage of time series that
 	// must fail the predicate for the
@@ -670,7 +670,7 @@ type TriggerParameters struct {
 	// that must fail the predicate for the
 	// condition to be triggered.
 	// +kubebuilder:validation:Optional
-	Count *int64 `json:"count,omitempty" tf:"count,omitempty"`
+	Count *float64 `json:"count,omitempty" tf:"count,omitempty"`
 
 	// The percentage of time series that
 	// must fail the predicate for the
